# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kashmir Local Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WorkerMembershipType = typeof WorkerMembershipType[keyof typeof WorkerMembershipType];


export const WorkerMembershipType = {
  free: 'free',
  premium: 'premium',
  premium_plus: 'premium_plus',
} as const;

export interface Worker {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  photoUrl?: string | null;
  profession: string;
  category: string;
  latitude: number;
  longitude: number;
  isAvailable: boolean;
  membershipType: WorkerMembershipType;
  averageRating: number;
  totalRatings: number;
  /** @nullable */
  area?: string | null;
  isApproved: boolean;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export type WorkerWithDistance = Worker & {
  distanceKm: number;
};

export interface WorkerInput {
  name: string;
  phone: string;
  photoUrl?: string;
  profession: string;
  category: string;
  latitude: number;
  longitude: number;
  area?: string;
  bio?: string;
}

export interface WorkerUpdate {
  name?: string;
  phone?: string;
  photoUrl?: string;
  profession?: string;
  category?: string;
  latitude?: number;
  longitude?: number;
  area?: string;
  bio?: string;
  isApproved?: boolean;
}

export interface AvailabilityUpdate {
  isAvailable: boolean;
}

export type MembershipUpdateMembershipType = typeof MembershipUpdateMembershipType[keyof typeof MembershipUpdateMembershipType];


export const MembershipUpdateMembershipType = {
  free: 'free',
  premium: 'premium',
  premium_plus: 'premium_plus',
} as const;

export interface MembershipUpdate {
  membershipType: MembershipUpdateMembershipType;
}

export interface Review {
  id: number;
  workerId: number;
  reviewerName: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  reviewerName: string;
  rating: number;
  comment?: string;
}

export interface Category {
  id: number;
  name: string;
  icon: string;
  slug: string;
}

export interface StatsSummary {
  totalWorkers: number;
  activeWorkers: number;
  premiumWorkers: number;
  premiumPlusWorkers: number;
  totalReviews: number;
  avgRating: number;
  pendingApprovals: number;
}

export interface CategoryStat {
  category: string;
  count: number;
  available: number;
}

export type ListWorkersParams = {
lat?: number;
lng?: number;
/**
 * Radius in km (default 20)
 */
radius?: number;
category?: string;
available?: boolean;
sort?: ListWorkersSort;
};

export type ListWorkersSort = typeof ListWorkersSort[keyof typeof ListWorkersSort];


export const ListWorkersSort = {
  distance: 'distance',
  rating: 'rating',
  membership: 'membership',
} as const;

export type GetNearbyWorkersParams = {
lat: number;
lng: number;
category?: string;
limit?: number;
};