# lib/api-zod/src/generated/types/categoryStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kashmir Local Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryStat {
  category: string;
  count: number;
  available: number;
}