# lib/api-zod/src/generated/types/membershipUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kashmir Local Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { MembershipUpdateMembershipType } from './membershipUpdateMembershipType';

export interface MembershipUpdate {
  membershipType: MembershipUpdateMembershipType;
}