# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kashmir Local Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  workerId: number;
  reviewerName: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}