# lib/api-zod/src/generated/types/workerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kashmir Local Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkerInput {
  name: string;
  phone: string;
  photoUrl?: string;
  profession: string;
  category: string;
  latitude: number;
  longitude: number;
  area?: string;
  bio?: string;
}