# lib/api-zod/src/generated/types/workerWithDistance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kashmir Local Services Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Worker } from './worker';

export type WorkerWithDistance = Worker & {
  distanceKm: number;
};